--- cypress.json
@@ -1,4 +1,0 @@
-{
-	"baseUrl": "http://localhost:3000",
-	"video": false
-}
