+++ docs/src/components/Nav.svelte
@@ -1,3 +1,4 @@
+
 <script>
 	export let segment;
 </script>
@@ -8,29 +9,24 @@
 		font-weight: 300;
 		padding: 0 1em;
 	}
-
 	ul {
 		margin: 0;
 		padding: 0;
 	}
-
 	/* clearfix */
 	ul::after {
 		content: '';
 		display: block;
 		clear: both;
 	}
-
 	li {
 		display: block;
 		float: left;
 	}
-
 	[aria-current] {
 		position: relative;
 		display: inline-block;
 	}
-
 	[aria-current]::after {
 		position: absolute;
 		content: '';
@@ -40,7 +36,6 @@
 		display: block;
 		bottom: -1px;
 	}
-
 	a {
 		text-decoration: none;
 		padding: 1em 0.5em;
