+++ docs/src/components/prism/import.js
@@ -1,0 +1,2 @@
+import prism from "prismjs";
+export default prism
