+++ docs/src/routes/_error.svelte
