+++ docs/src/routes/button/base.demo.svelte
@@ -1,0 +1,9 @@
+<script>
+  import Button from '@/components/Button.svelte'
+  let text = 'there'
+</script>
+
+<div>
+  <Button>Hello</Button>
+  <Button>{text}</Button>
+</div>
