+++ docs/src/routes/button/colored.demo.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Button from '../../../externals/components/Button.svelte'
+  import Button from '@/components/Button.svelte'
 </script>
 
 <div>
