+++ docs/src/routes/index.svelte
