+++ docs/src/service-worker.js
