+++ docs/static/favicon.png
