+++ docs/static/logo-192.png
