+++ docs/static/logo-512.png
