+++ docs/static/manifest.json
