--- src/components/prism/import.js
@@ -1,2 +1,0 @@
-import prism from "prismjs";
-export default prism
