--- src/routes/button/base.demo.svelte
@@ -1,9 +1,0 @@
-<script>
-  import Button from '../../../externals/components/Button.svelte'
-  let text = 'what!?'
-</script>
-
-<div>
-  <Button>Hello</Button>
-  <Button>{text}</Button>
-</div>
